refactor(tech): extract components to another path
