+++ packages/web/components/Technology/Details/Tabs.js
@@ -1,30 +1,10 @@
 import React from 'react';
-import PropTypes from 'prop-types';
-import styled from 'styled-components';
 import { resetIdCounter } from 'react-tabs';
-import { useTechnology } from '../../hooks';
-
-import { Tabs as Container, Tab, TabPanel, TabList } from '../../components/Tab';
-
-const TextValue = ({ title = '', value }) => {
-	return value ? (
-		<DescriptionValue>
-			{!!title && <strong>{title}: </strong>}
-			<span>{value}</span>
-		</DescriptionValue>
-	) : (
-		<></>
-	);
-};
-
-const Description = ({ title, children }) => (
-	<DescriptionContainer>
-		<TitleContainer>
-			<h4>{title}</h4>
-		</TitleContainer>
-		{children}
-	</DescriptionContainer>
-);
+import styled from 'styled-components';
+import { useTechnology } from '../../../hooks';
+import { Tab, TabList, TabPanel, Tabs as Container } from '../../Tab';
+import Description from './Description';
+import TextValue from './TextValue';
 
 const Tabs = () => {
 	const { technology } = useTechnology();
@@ -119,21 +99,6 @@ const Tabs = () => {
 	);
 };
 
-TextValue.propTypes = {
-	title: PropTypes.string,
-	value: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-};
-
-TextValue.defaultProps = {
-	title: '',
-	value: '',
-};
-
-Description.propTypes = {
-	title: PropTypes.string.isRequired,
-	children: PropTypes.node.isRequired,
-};
-
 Tabs.getInitialProps = () => {
 	resetIdCounter();
 };
@@ -173,46 +138,6 @@ export const Col = styled.div`
 
 	@media (max-width: ${({ theme }) => theme.screens.large}px) {
 		width: 100%;
-	}
-`;
-
-export const TitleContainer = styled.div`
-	border-bottom: 4px solid ${({ theme: { colors } }) => colors.primary};
-	width: calc(100% - 1rem);
-	margin: 2rem 0;
-
-	&:first-child {
-		margin-top: 0;
-	}
-
-	h4 {
-		display: inline-block;
-		background-color: ${({ theme: { colors } }) => colors.primary};
-		color: ${({ theme: { colors } }) => colors.white};
-		padding: 2.5rem 3rem;
-		font-weight: 600;
-		font-size: 1.8rem;
-		line-height: 1rem;
-		text-transform: uppercase;
-		margin-bottom: -4px;
-	}
-`;
-
-export const DescriptionContainer = styled.div`
-	margin-left: 1rem;
-	padding-bottom: 2rem;
-`;
-
-export const DescriptionValue = styled.p`
-	font-size: 1.6rem;
-	line-height: 2.4rem;
-
-	strong {
-		color: ${({ theme: { colors } }) => colors.darkGray};
-	}
-
-	span {
-		color: ${({ theme: { colors } }) => colors.black};
 	}
 `;
